+++ lib/src/widgets/raw_editor/raw_editor_state_text_input_client_mixin.dart
@@ -1,9 +1,12 @@
-import 'dart:ui';
+import 'dart:ui' show lerpDouble;
 
 import 'package:flutter/animation.dart' show Curves;
+import 'package:flutter/cupertino.dart' show CupertinoTheme;
 import 'package:flutter/foundation.dart' show kIsWeb;
+import 'package:flutter/material.dart' show Theme;
 import 'package:flutter/scheduler.dart' show SchedulerBinding;
 import 'package:flutter/services.dart';
+import 'package:flutter/widgets.dart' show MediaQuery;
 
 import '../../models/documents/document.dart';
 import '../../utils/delta.dart';
@@ -59,7 +62,10 @@ mixin RawEditorStateTextInputClientMixin on EditorState
           readOnly: widget.configurations.isReadOnly,
           inputAction: widget.configurations.textInputAction,
           enableSuggestions: !widget.configurations.isReadOnly,
-          keyboardAppearance: widget.configurations.keyboardAppearance,
+          keyboardAppearance: widget.configurations.keyboardAppearance ??
+              CupertinoTheme.maybeBrightnessOf(context) ??
+              MediaQuery.maybePlatformBrightnessOf(context) ??
+              Theme.of(context).brightness,
           textCapitalization: widget.configurations.textCapitalization,
           allowedMimeTypes:
               widget.configurations.contentInsertionConfiguration == null